Moved JSX, todoList array, to TodoList.js and also refactored App.js to use TodoList component. Verified that code is still working.
